+++ src/components/button/button_group/button_group.tsx
@@ -3,24 +3,24 @@ import classNames from 'classnames';
 
 import { EuiScreenReaderOnly } from '../../accessibility';
 import { ToggleType } from '../../toggle';
-import { IconType } from '../../icon';
 
 import { EuiButtonToggle } from '../button_toggle';
-import { Omit } from '../../common';
+import { Omit, CommonProps } from '../../common';
 
 import { ButtonColor } from '../button';
+import { IconType } from '../../icon';
 
 export interface EuiButtonGroupIdToSelectedMap {
   [id: string]: boolean;
 }
 
-export type GroupButtonSize = 's' | 'm';
+export type GroupButtonSize = 's' | 'm' | 'compressed';
 
-export interface EuiButtonGroupOption {
+export interface EuiButtonGroupOption extends CommonProps {
   id: string;
   label: string;
-  isDisabled?: boolean;
   name?: string;
+  isDisabled?: boolean;
   value?: any;
   iconSide?: 'left' | 'right';
   iconType?: IconType;
@@ -28,17 +28,21 @@ export interface EuiButtonGroupOption {
 
 export interface EuiButtonGroupProps {
   options?: EuiButtonGroupOption[];
-  onChange: (id: string, value: any) => void;
+  onChange: (id: string, value?: any) => void;
+  /**
+   * Typical sizing is `s`. Medium `m` size should be reserved for major features.
+   * `compressed` is meant to be used alongside and within compressed forms.
+   */
   buttonSize?: GroupButtonSize;
   isDisabled?: boolean;
   isFullWidth?: boolean;
   isIconOnly?: boolean;
   idSelected?: string;
-  idToSelectedMap?: EuiButtonGroupIdToSelectedMap;
   legend?: string;
   color?: ButtonColor;
-  type?: ToggleType;
   name?: string;
+  type?: ToggleType;
+  idToSelectedMap?: EuiButtonGroupIdToSelectedMap;
 }
 
 type Props = Omit<HTMLAttributes<HTMLDivElement>, 'onChange'> &
@@ -62,11 +66,16 @@ export const EuiButtonGroup: FunctionComponent<Props> = ({
 }) => {
   const classes = classNames(
     'euiButtonGroup',
+    [`euiButtonGroup--${buttonSize}`],
     {
       'euiButtonGroup--fullWidth': isFullWidth,
     },
     className
   );
+
+  const fieldsetClasses = classNames('euiButtonGroup__fieldset', {
+    'euiButtonGroup__fieldset--fullWidth': isFullWidth,
+  });
 
   let legendNode;
   if (legend) {
@@ -78,37 +87,56 @@ export const EuiButtonGroup: FunctionComponent<Props> = ({
   }
 
   return (
-    <fieldset>
+    <fieldset className={fieldsetClasses}>
       {legendNode}
 
       <div className={classes} {...rest}>
         {options.map((option, index) => {
+          const {
+            id,
+            name: optionName,
+            value,
+            isDisabled: optionDisabled,
+            className,
+            ...rest
+          } = option;
+
           let isSelectedState;
           if (type === 'multi') {
-            isSelectedState = idToSelectedMap[option.id] || false;
+            isSelectedState = idToSelectedMap[id] || false;
           } else {
-            isSelectedState = option.id === idSelected;
+            isSelectedState = id === idSelected;
           }
+
+          let fill;
+          if (buttonSize !== 'compressed') {
+            fill = isSelectedState;
+          }
+          const buttonClasses = classNames(
+            'euiButtonGroup__button',
+            {
+              'euiButtonGroup__button--selected': isSelectedState,
+            },
+            className
+          );
 
           return (
             <EuiButtonToggle
-              className="euiButtonGroup__button"
+              className={buttonClasses}
+              toggleClassName="euiButtonGroup__toggle"
+              id={id}
+              key={index}
+              value={value}
               color={color}
-              fill={isSelectedState}
-              iconSide={option.iconSide}
-              iconType={option.iconType}
-              id={option.id}
-              isDisabled={isDisabled || option.isDisabled}
+              fill={fill}
+              isDisabled={optionDisabled || isDisabled}
               isIconOnly={isIconOnly}
               isSelected={isSelectedState}
-              key={index}
-              label={option.label}
-              name={option.name || name}
-              onChange={() => onChange(option.id, option.value)}
-              size={buttonSize}
-              toggleClassName="euiButtonGroup__toggle"
+              name={optionName || name}
+              onChange={() => onChange(id, value)}
+              size={buttonSize === 'compressed' ? 's' : buttonSize}
               type={type}
-              value={option.value}
+              {...rest}
             />
           );
         })}